+++ releasenotes.rst
@@ -5,6 +5,7 @@ Dates are written as DD/MM/YYYY
 
 0.14.0 - unreleased
 -------------------
+- use enum.IntEnum/IntFlag instead of atom.IntEnum PR #453
 - fix operator bindings in template instances PR #445
 - fix FlowLayout error with FlowItems that have non-zero stretch or ortho_stretch PR #448
 - add support for styling notebook tabs PR #452